valve-server/requirements.txt
pyserial==3.5
tornado==6.4.1
